requirements.txt
torch
torchvision
customtkinter
tkinter
pillow